+++ src/test/unit/services/features/queryManagement/queryManagementService.test.ts
@@ -4,7 +4,7 @@ import {
   getCancelUrl,
   getCaption,
   getQueryManagement, removeSelectedDocument,
-  saveQueryManagement, uploadSelectedFile,
+  saveQueryManagement, updateQueryManagementDashboardItems, uploadSelectedFile,
 } from 'services/features/queryManagement/queryManagementService';
 import {QueryManagement, WhatDoYouWantToDo, WhatToDoTypeOption} from 'form/models/queryManagement/queryManagement';
 import express from 'express';
@@ -16,13 +16,95 @@ import {AppRequest} from 'models/AppRequest';
 import {CaseDocument} from 'models/document/caseDocument';
 import config from 'config';
 import nock from 'nock';
+import {DashboardTaskList} from 'models/dashboard/taskList/dashboardTaskList';
+import {Dashboard} from 'models/dashboard/dashboard';
+import {CaseRole} from 'form/models/caseRoles';
+import {DashboardTask} from 'models/dashboard/taskList/dashboardTask';
 
 jest.mock('../../../../../main/modules/draft-store/draftStoreService');
 jest.mock('../../../../../main/modules/utilityService');
 
+const mockExpectedDashboardInfo =
+  [{
+    'categoryEn': 'Hearing',
+    'categoryCy': 'Hearing Welsh',
+    tasks: [{
+      'id': '8c2712da-47ce-4050-bbee-650134a7b9e5',
+      'status': 'ACTION_NEEDED',
+      'taskNameEn': 'task_name_en',
+      'hintTextEn': 'hint_text_en',
+      'taskNameCy': 'task_name_cy',
+      'hintTextCy': 'hint_text_cy',
+    }, {
+      'id': '8c2712da-47ce-4050-bbee-650134a7b9e6',
+      'status': 'ACTION_NEEDED',
+      'taskNameEn': 'task_name_en',
+      'hintTextEn': 'hint_text_en',
+      'taskNameCy': 'task_name_cy',
+      'hintTextCy': 'hint_text_cy',
+    }],
+  }, {
+    'categoryEn': 'Applications',
+    'categoryCy': 'Applications Welsh',
+    tasks: [{
+      'id': '8c2712da-47ce-4050-bbee-650134a7b9e7',
+      'statusEn': 'ACTION_NEEDED',
+      'statusCy': 'ACTION_NEEDED',
+      'statusColour': 'govuk-red',
+      'taskNameEn': 'Contact the court to request a change to my case',
+      'hintTextEn': 'hint_text_en2',
+      'taskNameCy': 'task_name_cy2',
+      'hintTextCy': 'hint_text_cy2',
+    },
+    {
+      'id': '8c2712da-47ce-4050-bbee-650134a7b9e8',
+      'statusEn': 'ACTION_NEEDED',
+      'statusCy': 'ACTION_NEEDED',
+      'statusColour': 'govuk-red',
+      'taskNameEn': 'task_name_en2',
+      'hintTextEn': 'hint_text_en2',
+      'taskNameCy': 'task_name_cy2',
+      'hintTextCy': 'hint_text_cy2',
+    }] as DashboardTask[],
+  }] as DashboardTaskList[];
 const req = {params: {id: '123'}} as unknown as express.Request;
 const mockGetClaimById = utilityService.getClaimById as jest.Mock;
 
+describe('dashboard items update', () => {
+  const exclusionTask = new DashboardTaskList('Applications', 'Applications', []);
+  const dashboard = new Dashboard(mockExpectedDashboardInfo);
+  const claim = new Claim();
+
+  it('should update the header when exclusion matches', () => {
+    claim.caseRole = CaseRole.CLAIMANT;
+    updateQueryManagementDashboardItems(dashboard, exclusionTask, claim);
+
+    expect(dashboard.items[0].categoryEn).toEqual('Hearing');
+    expect(dashboard.items[1].categoryEn).toEqual('COMMON.QUERY_MANAGEMENT_DASHBOARD.APPLICATION_HEADING');
+  });
+
+  it('should update the tasks for GA and messages', () => {
+    claim.caseRole = CaseRole.CLAIMANT;
+    updateQueryManagementDashboardItems(dashboard, exclusionTask, claim);
+
+    expect(dashboard.items[1].tasks[0].taskNameEn).toEqual('COMMON.QUERY_MANAGEMENT_DASHBOARD.APPLICATIONS_TASK');
+    expect(dashboard.items[1].tasks[1].taskNameEn).toEqual('COMMON.QUERY_MANAGEMENT_DASHBOARD.VIEW_MESSAGES_TASK');
+  });
+
+  it('should update the task status if no queries exist for claimant', () => {
+    claim.caseRole = CaseRole.CLAIMANT;
+    updateQueryManagementDashboardItems(dashboard, exclusionTask, claim);
+
+    expect(dashboard.items[1].tasks[1].statusEn).toEqual('PAGES.TASK_LIST.NOT_AVAILABLE_YET');
+  });
+
+  it('should update the task status if no queries exist for defendant', () => {
+    claim.caseRole = CaseRole.DEFENDANT;
+    updateQueryManagementDashboardItems(dashboard, exclusionTask, claim);
+
+    expect(dashboard.items[1].tasks[1].statusEn).toEqual('PAGES.TASK_LIST.NOT_AVAILABLE_YET');
+  });
+});
 describe('save queryManagement data', () => {
   it('should save data successfully when query management not exists', async () => {
     //Given
@@ -176,7 +258,7 @@ describe('Uploading files', () => {
 
   beforeAll(() => {
     nock(civilServiceUrl).post('/case/document/generateAnyDoc').reply(200, returnedFile);
-  })
+  });
   beforeEach(() => {
     jest.resetAllMocks();
     const fileToUpload = {
@@ -185,7 +267,7 @@ describe('Uploading files', () => {
       mimetype: 'text/plain',
       size: 123,
       buffer: Buffer.from('test'),
-    }
+    };
     jest.spyOn(draftStoreService, 'generateRedisKeyForFile').mockReturnValue(mockKey);
     draftStoreGetFilesMock.mockResolvedValue([]);
     jest.spyOn(TypeOfDocumentSectionMapper, 'mapToSingleFile').mockReturnValue(fileToUpload);
@@ -223,7 +305,7 @@ describe('Uploading files', () => {
       mimetype: 'test',
       size: 123,
       buffer: Buffer.from('test'),
-    }
+    };
     jest.spyOn(TypeOfDocumentSectionMapper, 'mapToSingleFile').mockReturnValue(wrongMimeTypeFile);
 
     const saveSpy = jest.spyOn(draftStoreService, 'saveFilesToRedis');
